python/pyproject.toml
[project]
name = "python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.13",
    "pydantic>=2.11.7",
    "pypandoc-binary>=1.15",
    "requests>=2.32.4",
    "weasyprint>=65.1",
]
